Add tests for guest agent data
