+++ frontend/packages/kubevirt-plugin/integration-tests/tests/utils/constants/common.ts
@@ -24,6 +24,8 @@ export const VM_IMPORT_TIMEOUT_SECS = 160 * SEC;
 export const WINDOWS_IMPORT_TIMEOUT_SECS = 150 * SEC;
 export const VM_CREATE_AND_EDIT_TIMEOUT_SECS = 200 * SEC;
 export const VM_CREATE_AND_EDIT_AND_CLOUDINIT_TIMEOUT_SECS = 15 * 60 * SEC;
+export const VM_WITH_GA_CREATE_AND_EDIT_CLOUDINIT_TIMEOUT_SECS = 60 * 60 * SEC;
+export const GUEST_AGENT_FIELD_TIMEOUT_SECS = 30 * SEC;
 
 export const POD_CREATION_TIMEOUT_SECS = 40 * SEC;
 export const POD_TERMINATION_TIMEOUT_SECS = 30 * SEC;
