+++ frontend/packages/kubevirt-plugin/integration-tests/views/dashboard.view.ts
@@ -9,7 +9,7 @@ export const vmDetailsNode = vmDetails.get(4);
 export const vmDetailsIPAddress = vmDetails.get(5);
 export const vmDetailsOS = vmDetails.get(6);
 export const vmDetailsTZ = vmDetails.get(7);
-export const vmDetailsLoggedUser = vmDetails.get(8);
+export const vmDetailsLoggedInUsers = vmDetails.get(8);
 export const vmDetailsViewAll = $$('.co-dashboard-card__link').get(0);
 
 export const vmStatus = $('.co-status-card__health-body > span');
