+++ frontend/packages/kubevirt-plugin/integration-tests/views/vm.disks.view.ts
@@ -1,0 +1,7 @@
+import { $ } from 'protractor';
+
+export const fileSystemsTableHeader = $('#file-systems-header');
+export const fileSystemsTable = $(`[aria-label="FileSystems"]`)
+  .$('div')
+  .$('table')
+  .$('tbody');
